+++ tests/po_lib/__init__.py
@@ -5,6 +5,7 @@ from typing import Dict, Any, Callable
 
 from url_matcher import Patterns
 
+from .. import po_sub_lib  # NOTE: this module contains a PO with @handle_rules
 from web_poet import handle_urls, PageObjectRegistry
 
 
@@ -22,7 +23,7 @@ class POTopLevelOverriden2:
     ...
 
 
-secondary_registry = PageObjectRegistry(name="secondary")
+secondary_registry = PageObjectRegistry()
 
 
 # This first annotation is ignored. A single annotation per registry is allowed