+++ CHANGELOG.rst
@@ -5,8 +5,9 @@ Changelog
 TBR
 ------------------
 
-* ``handle_urls`` decorator and ``find_page_object_overrides`` function added.
-* new CLI tool for displaying all available Page Objects: ``web_poet <path>``
+* added a ``PageObjectRegistry`` class which has the  ``handle_urls`` decorator
+  to write override rules.
+* new CLI tool for displaying all available Page Objects: ``web_poet <module>``
 * removed support for Python 3.6
 * added support for Python 3.10